sessions mm.

+++ src/main/java/com/wishlist/Controller/UserController.java
@@ -3,6 +3,8 @@ package com.wishlist.Controller;
 import com.wishlist.Model.*;
 import com.wishlist.Service.UserService;
 import com.wishlist.Service.WishlistService;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpSession;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
@@ -19,6 +21,7 @@ public class UserController {
         this.userService = userService;
         this.wishlistService = wishlistService;
     }
+
     @GetMapping("/create")
     public String showCreateForm(Model model) {
         User newUser = new User();
@@ -40,14 +43,35 @@ public class UserController {
     }
 
     @PostMapping("/login")
-    public String loginUser(@ModelAttribute User user) {
-        int userID = userService.getUserByUsername(user.getUsername()).getID();
-        userService.login(user.getUsername(), user.getPassword());
-        return "redirect:/user/" + userID;
+    public String loginUser(@ModelAttribute User user, HttpSession session, Model model) {
+        User dbUser = userService.getUserByUsername(user.getUsername());
+
+        if (dbUser != null && dbUser.getPassword().equals(user.getPassword())) {
+            session.setAttribute("userID", dbUser.getID());
+            session.setMaxInactiveInterval(30 * 60);
+            return "redirect:/user/" + dbUser.getID();
+        } else {
+            model.addAttribute("wrongCredentials", true);
+            return "login-user";
+        }
+    }
+    @GetMapping("/logout")
+    public String logout(HttpSession session, HttpServletRequest request) {
+        session.invalidate();
+        request.getSession(true);
+        return "redirect:/user/login";
     }
 
+
+
     @GetMapping("/{uid}")
-    public String userPage(Model model, @PathVariable int uid) {
+    public String userPage(Model model, @PathVariable int uid, HttpSession session) {
+        Integer userId = (Integer) session.getAttribute("userID");
+        if (userId == null || !userId.equals(uid)) {
+            return "redirect:/user/login";
+        }
+
+
         User userByID = userService.getUserByID(uid);
         List<Wishlist> userWishlists = wishlistService.getAllWishlistsByUserWithID(uid);
 
@@ -56,31 +80,49 @@ public class UserController {
         return "userByID";
     }
 
-    @GetMapping("/{uid}/update")
-    public String updateUser(Model model, @PathVariable(value="uid") int id) {
-        User updateUser = userService.getUserByID(id);
 
+    @GetMapping("/{uid}/update")
+    public String updateUser(Model model, @PathVariable int uid, HttpSession session) {
+        Integer userId = (Integer) session.getAttribute("userID");
+        if (userId == null || !userId.equals(uid)) {
+            return "redirect:/user/login";
+        }
+
+
+        User updateUser = userService.getUserByID(uid);
         User userDTO = new User();
         userDTO.setName(updateUser.getName());
         userDTO.setUsername(updateUser.getUsername());
         userDTO.setPassword(updateUser.getPassword());
 
-        model.addAttribute("ID", id);
+        model.addAttribute("ID", uid);
         model.addAttribute("userDTO", userDTO);
         return "update-user";
     }
 
     @PostMapping("/{uid}/update")
-    public String updateUser(@ModelAttribute User userDTO, @PathVariable(value="uid") int id) {
+    public String updateUser(@ModelAttribute User userDTO, @PathVariable int uid, HttpSession session) {
+        Integer userId = (Integer) session.getAttribute("userID");
+        if (userId == null || !userId.equals(uid)) {
+            return "redirect:/user/login";
+        }
 
-        User updateUser = new User(id, userDTO.getName(), userDTO.getUsername(), userDTO.getPassword());
+
+        User updateUser = new User(uid, userDTO.getName(), userDTO.getUsername(), userDTO.getPassword());
         userService.updateUser(updateUser);
-        return "redirect:/user/login";
+        return "redirect:/user/" + uid;
     }
 
     @PostMapping("/{uid}/delete")
-    public String deleteUser(@PathVariable int uid) {
+    public String deleteUser(@PathVariable int uid, HttpSession session) {
+        Integer userId = (Integer) session.getAttribute("userID");
+        if (userId == null || !userId.equals(uid)) {
+            return "redirect:/user/login";
+        }
+
+
         userService.deleteUserByID(uid);
+        session.invalidate();
         return "redirect:/";
     }
 }
